Cross Site Request Forgery
in the project was enable and disable the csrf protection with the end to understand when is useful to use this protection. the recommendation to use CSRF is for any request that could be processed by a browser by anormal user,  otherwise if you are only creating a service that is used by non-browser clients, you will likely want to disable CSRF protection.